requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
psycopg2-binary>=2.9.0,<3.0
django-environ>=0.11.2,<1.0
celery>=5.3.6,<6.0
redis>=5.0.0,<6.0
python-dotenv>=1.0.0,<2.0
gunicorn>=21.2.0,<22.0
django-cors-headers>=4.0.0,<5.0

